fix(editor,desktop): clear fallow dead-code (unused theme exports, atoms<->attachments cycle)
The Fallow Advisory Envelopes job has been red on this PR all along: the
dead-code lane reports findings un-attributed (so envelope `introduced: 0` is not
"not mine"). All 3 findings are this PR's own code:

- WorkbenchThemeProvider: greenWorkbenchThemeOptions and workbenchTheme were
  exported but used only internally — un-export them.
- editor/chat: break the atoms.ts <-> attachments.tsx circular import by
  extracting the pure ComposerAttachment model (schema, capture validation,
  object-URL helpers; no atom/React deps) into attachment-model.ts. atoms.ts now
  imports the model from there, leaving attachments.tsx (plugins/UI) as the only
  edge.

`beep quality fallow dead-code --check` now reports 0 findings (exit 0).
